+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/test/StoreLoadTest.java
@@ -68,8 +68,8 @@ public class StoreLoadTest {
 		}
 
 		private static void selectRandomRow(EntityTableModel<?> tableModel) {
-			if (tableModel.getRowCount() > 0) {
-				tableModel.selectionModel().setSelectedIndex(RANDOM.nextInt(tableModel.getRowCount()));
+			if (tableModel.rowCount() > 0) {
+				tableModel.selectionModel().setSelectedIndex(RANDOM.nextInt(tableModel.rowCount()));
 			}
 		}
 	}