+++ demos/petstore/src/main/java/is/codion/framework/demos/petstore/testing/PetstoreLoadTest.java
@@ -78,8 +78,8 @@ public final class PetstoreLoadTest {
 		}
 
 		private static void selectRandomRow(EntityTableModel<?> tableModel) {
-			if (tableModel.getRowCount() > 0) {
-				tableModel.selectionModel().setSelectedIndex(RANDOM.nextInt(tableModel.getRowCount()));
+			if (tableModel.rowCount() > 0) {
+				tableModel.selectionModel().setSelectedIndex(RANDOM.nextInt(tableModel.rowCount()));
 			}
 		}
 	}